4.DevOps_phase/ltc-devops-project/api/requirements.txt
boto3
fastapi
uvicorn
pydantic
starlette
python-dotenv
